TFJ-790 getRetweetCount now returns int

+++ twitter4j-appengine/src/main/java/twitter4j/internal/json/LazyStatus.java
@@ -234,7 +234,7 @@ final class LazyStatus implements twitter4j.Status {
      *
      * @return the retweet count.
      */
-    public long getRetweetCount() {
+    public int getRetweetCount() {
         return getTarget().getRetweetCount();
     }
